requirements.txt
appdirs==1.4.4
Babel==2.9.0
bce-python-sdk==0.8.53
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Babel==2.0.0
future==0.18.2
identify==1.5.14
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lmdb==1.1.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
networkx==2.5
nodeenv==1.5.0
numpy==1.20.1
opencv-python==4.2.0.32
paddleocr==2.0.2
Pillow==8.1.0
pre-commit==2.10.1
protobuf==3.15.2
pyclipper==1.2.1
pycodestyle==2.6.0
pycryptodome==3.10.1
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.1
Shapely==1.7.1
shellcheck-py==0.7.1.1
six==1.15.0
tifffile==2021.2.1
toml==0.10.2
tqdm==4.57.0
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
visualdl==2.1.1
Werkzeug==1.0.1
zipp==3.4.0
